third_party/blink/web_tests/flag-specific/disable-layout-ng/svg/dynamic-updates/SVGTextElement-svgdom-lengthAdjust-prop-expected.txt

SVG 1.1 dynamic update tests
Stretched text

Tests dynamic updates of the 'lengthAdjust' property of the SVGTextElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS textElement.getAttribute('lengthAdjust') is null
PASS textElement.lengthAdjust.baseVal is SVGTextContentElement.LENGTHADJUST_SPACING
PASS textElement.textLength.baseVal.value is 200
PASS lastLength = textElement.getComputedTextLength(); lastLength > 0 && lastLength < 200 is true
PASS textElement.getAttribute('lengthAdjust') is "spacingAndGlyphs"
PASS textElement.lengthAdjust.baseVal is SVGTextContentElement.LENGTHADJUST_SPACINGANDGLYPHS
PASS textElement.textLength.baseVal.value is 200
PASS textElement.getComputedTextLength() == lastLength is true
PASS successfullyParsed is true

TEST COMPLETE
